requirements.txt
python-dotenv==1.1.1
colorama==0.4.6
